Move Lazy Loading into CodeGenerationManager and remove LazyServiceConnection.
Fixes #37
