+++ blocklylib-core/src/main/java/com/google/blockly/android/AbstractBlocklyActivity.java
@@ -16,20 +16,18 @@
 package com.google.blockly.android;
 
 import android.content.Context;
-import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.res.AssetManager;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v4.app.Fragment;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.Menu;
@@ -41,20 +39,20 @@ import android.widget.Toast;
 
 import com.google.blockly.android.clipboard.BlockClipDataHelper;
 import com.google.blockly.android.clipboard.SingleMimeTypeClipDataHelper;
-import com.google.blockly.android.codegen.CodeGeneratorService;
+import com.google.blockly.android.codegen.CodeGenerationRequest;
 import com.google.blockly.android.codegen.CodeGeneratorManager;
 import com.google.blockly.android.control.BlocklyController;
 import com.google.blockly.android.ui.BlockViewFactory;
+import com.google.blockly.android.ui.BlocklyUnifiedWorkspace;
 import com.google.blockly.android.ui.DeleteVariableDialog;
 import com.google.blockly.android.ui.NameVariableDialog;
 import com.google.blockly.android.ui.TrashCanView;
 import com.google.blockly.android.ui.WorkspaceHelper;
-import com.google.blockly.android.ui.BlocklyUnifiedWorkspace;
-import com.google.blockly.android.codegen.CodeGenerationRequest;
 import com.google.blockly.model.Block;
 import com.google.blockly.model.BlockFactory;
 import com.google.blockly.model.BlocklySerializerException;
 import com.google.blockly.model.Workspace;
+import com.google.blockly.utils.StringOutputStream;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -747,11 +745,27 @@ public abstract class AbstractBlocklyActivity extends AppCompatActivity {
      * @see #getCodeGenerationCallback()
      */
     protected void onRunCode() {
-        mCodeGeneratorManager.requestCodeGeneration(
-            getBlockDefinitionsJsonPaths(),
-            getGeneratorsJsPaths(),
-            mWorkspaceFragment.getWorkspace(),
-            getCodeGenerationCallback());
+        Workspace workspace = mWorkspaceFragment.getWorkspace();
+        if (workspace.hasBlocks()) {
+            try {
+                final StringOutputStream serialized = new StringOutputStream();
+                workspace.serializeToXml(serialized);
+                CodeGenerationRequest codeGenerationRequest =
+                    new CodeGenerationRequest(
+                        serialized.toString(),
+                        getCodeGenerationCallback(),
+                        getBlockDefinitionsJsonPaths(),
+                        getGeneratorsJsPaths());
+                mCodeGeneratorManager.requestCodeGeneration(codeGenerationRequest);
+            } catch (BlocklySerializerException e) {
+                Log.wtf(TAG, e);
+                Toast.makeText(this, "Code generation failed.",
+                    Toast.LENGTH_LONG).show();
+
+            }
+        } else {
+            Log.i(TAG, "No blocks in workspace. Skipping run request.");
+        }
     }
 
     /**
